+++ client/components/auth-form.js
@@ -64,7 +64,6 @@ const mapDispatch = dispatch => {
       const formName = evt.target.name
       const email = evt.target.email.value
       const password = evt.target.password.value
-      console.log(`Running auth thunk....`)
       dispatch(auth(email, password, formName))
     }
   }
