+++ client/store/singleUser.js
@@ -2,6 +2,7 @@ import axios from 'axios'
 
 //Action Types
 const GET_ONE_USER = `GET_ONE_USER`
+const UPDATE_USER = 'UPDATE_USER'
 //Action Creators
 const getOneUser = user => {
   return {
@@ -9,6 +10,7 @@ const getOneUser = user => {
     user
   }
 }
+const updateUser = updatedUser => ({type: UPDATE_USER, updatedUser})
 
 //Thunk Creators
 export const getOneUserThunk = userID => {
@@ -22,12 +24,25 @@ export const getOneUserThunk = userID => {
   }
 }
 
+export const updateUserThunk = (userID, updatedUser) => {
+  return async dispatch => {
+    try {
+      const {data} = await axios.put(`/api/users/${userID}`, updatedUser)
+      console.log('data-------------', data)
+      dispatch(updateUser(data[1]))
+    } catch (err) {
+      console.error(err)
+    }
+  }
+}
 //Reducer
 const initState = {}
 export default function(state = initState, action) {
   switch (action.type) {
     case GET_ONE_USER:
       return action.user
+    case UPDATE_USER:
+      return action.updatedUser
     default:
       return state
   }
